@@ -1,6 +1,4 @@
 ﻿using System.Linq;
-using System.Runtime.InteropServices.ComTypes;
-using System.Security.Cryptography.X509Certificates;
 using Core.BLL.ViewModels;
 using Core.DAL.Entities;
 using AutoMapper;
@@ -34,29 +32,18 @@ namespace Core.BLL.Mappers
 				.ForMember(x => x.CreatedBy, x => x.MapFrom(y => y.CreatedBy))
 				.ForMember(x => x.StartsOn, x => x.MapFrom(y => y.StartsOn))
 				.ForMember(x => x.EndsOn, x => x.MapFrom(y => y.EndsOn))
-				.ForMember(
-					x => x.Experiences,
-					x => 
-						x.MapFrom(y => y.Experiences));
+				.ForMember(x => x.Experiences, x => x.Ignore())
+				.ForMember(x => x.Participants, x => x.Ignore());
 			CreateMap<AdventureViewModel, Adventure>()
 				.ForMember(x => x.Id, x => x.MapFrom(y => y.Id))
 				.ForMember(x => x.Name, x => x.MapFrom(y => y.Name))
 				.ForMember(x => x.Description, x => x.MapFrom(y => y.Description))
-				.ForMember(x => x.CreatedBy, x => x.MapFrom(y => y.CreatedBy))
+				.ForMember(x => x.CreatedBy, x => x.Ignore())
+				.ForMember(x => x.CreatedById, x => x.Ignore())
 				.ForMember(x => x.StartsOn, x => x.MapFrom(y => y.StartsOn))
 				.ForMember(x => x.EndsOn, x => x.MapFrom(y => y.EndsOn))
-				.ForMember(
-					x => x.Experiences,
-					x =>
-						x.MapFrom(y => y.Experiences
-							.Select(Mapper.Map<Experience>)
-							.ToList()));
-				//.ForMember(
-				//	x => x.AdventureUsers,
-				//	x => 
-				//		x.MapFrom(y => y.Participants
-				//			.Select(Mapper.Map<User>)
-				//			.ToList()));
+				.ForMember(x => x.Experiences, x => x.Ignore())
+				.ForMember(x => x.AdventureUsers, x => x.Ignore());
 
 			CreateMap<Experience, ExperienceViewModel>()
 				.ForMember(x => x.Id, x => x.MapFrom(y => y.Id))
