@@ -1,4 +1,6 @@
-﻿using Core.BLL.Interfaces;
+﻿using System;
+using System.Linq;
+using Core.BLL.Interfaces;
 using Core.BLL.ViewModels;
 using Core.DAL.Entities;
 using Core.DAL.Interfaces;
@@ -11,6 +13,23 @@ namespace Core.BLL.Services
 		public AdventureService(IAdventureRepository adventureRepository, IMapper mapper)
 			: base(adventureRepository, mapper)
 		{
+		}
+
+		public override int Create(AdventureViewModel viewModel)
+		{
+            var entity = Mapper.Map<Adventure>(viewModel);
+
+            entity.Experiences = viewModel.Experiences
+                .Select(x => Mapper.Map<Experience>(x))
+                .ToList();
+            entity.AdventureUsers = viewModel.Participants
+                .Select(x => Mapper.Map<AdventureUser>(x))
+                .ToList();
+            entity.CreatedBy = Mapper.Map<User>(viewModel.CreatedBy);
+
+            var id = Repository.Create(entity);
+
+			return id;
 		}
 	}
 }
