@@ -5,6 +5,9 @@ namespace Core.DAL
 {
 	public class CoreDbContext : DbContext
 	{
+		private const int defaultStartValue = 100;
+		private const string defaultSchema = "public";
+
 		public CoreDbContext(DbContextOptions<CoreDbContext> options)
 			: base(options)
 		{
@@ -19,15 +22,34 @@ namespace Core.DAL
 
 		protected override void OnModelCreating(ModelBuilder modelBuilder)
 		{
-			modelBuilder.HasDefaultSchema("public");
-			modelBuilder.ForNpgsqlUseIdentityAlwaysColumns();
-			modelBuilder.ForNpgsqlUseSequenceHiLo();
+			modelBuilder.HasDefaultSchema(defaultSchema);
+			
+			modelBuilder.UseIdentityColumns();
+			modelBuilder.UseIdentityByDefaultColumns();
+
+			modelBuilder.Entity<User>().Property(b => b.Id).HasIdentityOptions(startValue: defaultStartValue);
+			modelBuilder.Entity<Location>().Property(b => b.Id).HasIdentityOptions(startValue: defaultStartValue);
+			modelBuilder.Entity<Experience>().Property(b => b.Id).HasIdentityOptions(startValue: defaultStartValue);
+			modelBuilder.Entity<Adventure>().Property(b => b.Id).HasIdentityOptions(startValue: defaultStartValue);
+			modelBuilder.Entity<AdventureUser>().Property(b => b.Id).HasIdentityOptions(startValue: defaultStartValue);
 
 			modelBuilder.Entity<User>()
-				.HasIndex(b => b.Name)
+                .HasIndex(b => b.Name)
+                .IsUnique();
+
+			modelBuilder.Entity<Location>()
+				.HasIndex(x => x.Name)
 				.IsUnique();
 
-			base.OnModelCreating(modelBuilder);
+			modelBuilder.Entity<Experience>()
+				.HasIndex(x => x.Name)
+				.IsUnique();
+
+			modelBuilder.Entity<Adventure>()
+				.HasIndex(x => x.Name)
+				.IsUnique();
+
+            base.OnModelCreating(modelBuilder);
 		}
 
 		public override int SaveChanges()
