@@ -6,8 +6,8 @@ namespace Core.DAL.Entities
 	public class BaseEntity
 	{
 		[Key]
+		[Column("Id", Order = 1, TypeName = "int")]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		[Column("Id", Order = 1, TypeName = "integer")]
 		public int? Id { get; set; }
 	}
 }
