@@ -5,7 +5,7 @@ class BaseService {
 	constructor(path) {
 		this.basePath = `/api/${path}`;
 		this.api = axios.create({
-			baseURL: 'http://localhost:3000',
+			baseURL: 'http://localhost:5000',
 			headers: {
 				'Content-Type': 'application/json'
 			}
